fix compilation warning for read_only check
Recent compilers produce a warning about the ternary operator shortcut:
==============
peekpoke.c:316:58: warning: the omitted middle operand in "?:" will
always be "true", suggest explicit middle operand [-Wparentheses]
==============

Looking closer this expression was actually wrong: if read_only is set,
we want to OR in 0, not read_only. This just happened to work, because
PROT_READ is 1, the same as read_only if it's true.
Since shortcutting the ternary operator in this way is a GNU C extension
anyway, let's just be explicit here and OR in the 0.

Reported-by: Github user @rbalykov