properly handle base address
The -b option to set a base address was overly simply implemented: it
just mapped one page, hoping that every address would fit in there.
An alternative would be to just map everthing till the highest address
used in any command. But that might not be feasible, as there could be
potentially large gaps in between, which we might not be allowed to map.
This is especially true when no offset is given at all, so absolute
addresses are specified for every command.

Fix this more properly by adding a (resizable) array holding mapping
ranges. We use the check_command phase to first build this array, then
map the required pages (and just these) before we start executing the
commands.

A potential extension would be to support multiple -b options, to change
the base address between commands. This is not yet implemented.