add "dump range" support
Add a "-d" and "-D" option to dump the content of a range of memory, in
a hexdump style manner. The second argument after the address is the
number of words to dump, so "d.l 0x0 9" would dump 36 bytes in total.
"-d" shows bytes (like hexdump -C):
00000000  78 56 34 12 32 54 76 98
"-D" shows the data in word size:
00000000  12345678 98765432